* defobj.h (types_t): Make _long_double of that type.
* internal.m (typeError): New function.
(CONVERTNUMBER, ENSUREVALUETYPE): New macros.  Latter uses typeError.
(object_setVariableFromExpr): Use for former all numerics and
latter for object, class, boolean, and char


git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@6146 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/src/defobj/defobj.h
@@ -1045,7 +1045,7 @@ typedef union {
   unsigned long long ulonglong;
   float _float;
   double _double;
-  double _long_double;
+  long double _long_double;
 } types_t;
 
 typedef void *JOBJECT;
